% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{as.best.data.frame}
\alias{as.best.data.frame}
\title{Convert Dataframe Columns to Best of Numeric or Character}
\usage{
\method{as.best}{data.frame}(x, prefix = "#", ...)
}
\arguments{
\item{x}{data.frame}

\item{prefix}{character to prepend to values in mixed numeric/character columns}

\item{...}{passed to other methods}
}
\value{
data.frame
}
\description{
Convert data.frame columns to best of numeric or character.
}
\details{
Converts columns to numeric if doing so creates no new NA; otherwise to character. Number-like columns that are nevertheless character are prefixed by default to make this explicit when viewing only a few rows.
}

