% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{as.best.default}
\alias{as.best.default}
\title{Convert Vector to Best of Numeric or Character}
\usage{
\method{as.best}{default}(x, prefix = "", na.strings = c(".", "NA", ""),
  ...)
}
\arguments{
\item{x}{vector}

\item{prefix}{character to prepend to values in mixed numeric/character columns}

\item{...}{passed to other methods}
}
\value{
vector
}
\description{
Convert vector to best of numeric or character.
}
\details{
Converts vector to numeric if doing so creates no new NA; otherwise to character. Number-like vectors that are nevertheless character are prefixed by default to make this explicit when viewing only a few rows.
}

