% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{as.best}
\alias{as.best}
\title{Convert to Best of Numeric or Character}
\usage{
as.best(x, ...)
}
\arguments{
\item{x}{vector or data.frame}

\item{...}{passed to other methods}
}
\description{
Convert to best of numeric or character.
}
\details{
Converts to numeric if doing so creates no new NA; otherwise to character.
}

