% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{as.bootstrap.filepath}
\alias{as.bootstrap.filepath}
\title{Create a Bootstrap Table from Filepath}
\usage{
\method{as.bootstrap}{filepath}(x, skip = 28, check.names = FALSE,
  lo = "2.5", hi = "97.5", verbose = TRUE, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{skip}{number of lines to skip in bootstrap_results.csv}

\item{check.names}{passed to bootstrap reader}

\item{lo}{the PsN bootstrap lower confidence limit (\%)}

\item{hi}{the PsN bootstrap upper confidence limit (\%)}

\item{verbose}{display messages}

\item{...}{arguments to methods}
}
\value{
data.frame
}
\description{
Creates a bootstrap table from a PsN bootstrap results csv filepath.
}

