% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definitions.R
\name{as.definitions.modelname}
\alias{as.definitions.modelname}
\alias{definitions}
\title{Create a Item Definitions from Model Name}
\usage{
\method{as.definitions}{modelname}(x, verbose = FALSE,
  opt = getOption("project"), project = if (is.null(opt)) getwd() else opt,
  rundir = file.path(project, x), ctlfile = file.path(rundir, paste0(x,
  ".ctl")), fields = c("symbol", "label", "unit"), unique = TRUE, ...)
}
\arguments{
\item{x}{a model name (numeric or character)}

\item{verbose}{set FALSE to suppress messages}

\item{opt}{alternative argument for setting project}

\item{project}{parent directory of model directories}

\item{rundir}{specific model directory}

\item{ctlfile}{path to control stream}

\item{fields}{metadata fields to read from control stream if no metafile}

\item{...}{passed to other functions}

\item{lo}{the PsN bootstrap lower confidence limit (\%)}

\item{hi}{the PsN bootstrap upper confidence limit (\%)}

\item{metafile}{metadata for item definitions; will be created if missing (edit and re-run)}

\item{xmlfile}{path to xml file}

\item{bootcsv}{path to PsN bootstrap_results.csv}

\item{strip.namespace}{whether to strip e.g. nm: from xml elements for easier xpath syntax}

\item{skip}{number of lines to skip in bootstrap_results.csv}

\item{check.names}{passed to bootstrap reader}

\item{digits}{limits numerics to significant digits if specified}

\item{ci}{combine bootstrap lo and hi into an enclosed interval}

\item{sep}{separator for bootstrap interval}

\item{open}{first character for bootstrap interval}

\item{close}{last character for bootstrap interval}

\item{format}{format numerics as character}

\item{relative}{transform standard errors to relative standard errors: rse replaces se}

\item{percent}{if relative is true, express as percent (else ignore): prse replaces se}

\item{quote}{quote csv values when creating template metafile}

\item{na}{how to encode NA in metafile}

\item{na.strings}{what to read as NA in metafile}

\item{as.is}{passed to read.csv for reading metafile}

\item{nonzero}{limit random effects to those with nonzero estimates}
}
\value{
object of class definitions, data.frame
}
\description{
Creates a item definitions from a model name. Pass the project argument or set 
the project option.
}
\details{
Normally you can just call the generic.  Suitable defaults are supplied, but much customization is supported by means of arguments documented here and in called functions.
}
\examples{
library(magrittr)
options(project = system.file('project/model',package='partab'))
1001 \%>\% as.definitions
}
\seealso{
\code{\link{as.docx.definitions}}

\code{\link{as.flextable.definitions}}

\code{\link{as.xml_document.modelname}}

\code{\link{as.bootstrap.modelname}}

\code{\link{as.nmctl.modelname}}

\code{\link{write.csv}}

\code{\link{read.csv}}
}

