% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{as.docx.data.frame}
\alias{as.docx.data.frame}
\title{Coerce data.frame to docx}
\usage{
\method{as.docx}{data.frame}(x, title = "untitled", template,
  empty_template = FALSE,
  list.definition = getOption("ReporteRs-list-definition"), ...)
}
\arguments{
\item{x}{docx}

\item{title}{passed to docx}

\item{template}{passed to docx}

\item{empty_template}{passed to docx}

\item{list.definition}{passed to docx}

\item{...}{passed to addFlexTable}
}
\value{
docx
}
\description{
Coerces data.frame to docx
}
\examples{
library(magrittr)
options(project = system.file('project/model',package='partab'))
1001 \%>\% as.partab \%>\% as.docx \%>\% as.file('1001.docx')
}
\seealso{
\code{\link{ReporteRs::docx}}

\code{\link{ReporteRs::addFlexTable}}

\code{\link{ReporteRs::FlexTable}}
}

