% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{as.docx.docx}
\alias{as.docx.docx}
\title{Coerce docx to docx}
\usage{
\method{as.docx}{docx}(x, ...)
}
\arguments{
\item{x}{docx}

\item{...}{dots}
}
\value{
docx
}
\description{
Coerces docx to docx.
}
\details{
Returns the object.
}

