% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{as.docx}
\alias{as.docx}
\title{Coerce to docx}
\usage{
as.docx(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{arguments to methods}
}
\value{
docx
}
\description{
Coerces to docx
}
\seealso{
\code{link{as.docx.data.frame}}
}

