% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{as.file.docx}
\alias{as.file.docx}
\title{Coerce docx to File}
\usage{
\method{as.file}{docx}(x, file, ...)
}
\arguments{
\item{x}{docx}

\item{file}{storage path for docx file}

\item{...}{passed to writeDoc}
}
\description{
Coerces docx to file.
}
\seealso{
\code{\link[ReporteRs]{docx}}

\code{\link[ReporteRs]{writeDoc}}
}

