% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{as.flextable.data.frame}
\alias{as.flextable.data.frame}
\title{Coerce data.frame to flextable}
\usage{
\method{as.flextable}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{passed to flextable if valid argument}
}
\value{
flextable
}
\description{
Coerces data.frame to flextable
}
\seealso{
\code{\link{ReporteRs::docx}}

\code{\link{ReporteRs::addFlexTable}}

\code{\link{ReporteRs::FlexTable}}
}

