% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{as.flextable.partab}
\alias{as.flextable.partab}
\title{Coerce partab to flextable}
\usage{
\method{as.flextable}{partab}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{passed to flextable if valid argument}
}
\value{
flextable
}
\description{
Coerces partab to flextable.
}
\details{
At present, just calls the data.frame method.
}
\seealso{
\code{\link[ReporteRs]{docx}}

\code{\link[ReporteRs]{addFlexTable}}

\code{\link[ReporteRs]{FlexTable}}
}

