% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{as.initList.character}
\alias{as.initList.character}
\title{Coerce character to initList}
\usage{
\method{as.initList}{character}(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{dots}
}
\value{
initList
}
\description{
Coerces character to initList.
}

