% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{as.initList.initList}
\alias{as.initList.initList}
\title{Coerce initList to initList}
\usage{
\method{as.initList}{initList}(x, ...)
}
\arguments{
\item{x}{initList}

\item{...}{dots}
}
\value{
initList
}
\description{
Coerces initList to initList
}

