% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{as.initList.list}
\alias{as.initList.list}
\title{Coerce list to initList}
\usage{
\method{as.initList}{list}(x, comment = character(0), ...)
}
\arguments{
\item{x}{list}

\item{comment}{character}

\item{...}{dots}
}
\value{
initList
}
\description{
Coerces list to initlist.
}

