% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{as.initList.numeric}
\alias{as.initList.numeric}
\title{Coerce numeric to initList}
\usage{
\method{as.initList}{numeric}(x, fixed = FALSE, comment = character(0), ...)
}
\arguments{
\item{x}{numeric}

\item{fixed}{logical}

\item{comment}{character}

\item{...}{dots}
}
\value{
initList
}
\description{
Coerces numeric to initList
}

