% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{as.nmctl.character}
\alias{as.nmctl.character}
\title{Coerce character to nmctl
Coerces chacter to nmctl.}
\usage{
\method{as.nmctl}{character}(x, pattern = "^ *\\\\$([^ ]+)( .*)?$",
  head = "\\\\1", tail = "\\\\2", parse = FALSE, ...)
}
\arguments{
\item{x}{character}

\item{pattern}{pattern to identify record declarations}

\item{head}{subpattern to identify declaration type}

\item{tail}{subpattern remaining}

\item{parse}{whether to convert thetas to initList objects}

\item{...}{dots}
}
\value{
list
}
\description{
Coerce character to nmctl
Coerces chacter to nmctl.
}

