% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{as.nmctl.filename}
\alias{as.nmctl.filename}
\title{Convert Filename to nmctl}
\usage{
\method{as.nmctl}{filename}(x, parse, ...)
}
\arguments{
\item{x}{filename}

\item{parse}{convert thetas to initList}

\item{...}{dots}
}
\value{
nmctl
}
\description{
Converts filename to nmctl.
}

