% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{as.nmctl.modelname}
\alias{as.nmctl.modelname}
\title{Convert Modelname to nmctl}
\usage{
\method{as.nmctl}{modelname}(x, verbose = TRUE, project = if (is.null(opt))
  getwd() else opt, opt = getOption("project"), rundir = file.path(project,
  x), ctlfile = file.path(rundir, paste0(x, ext)), ext = ".ctl",
  parse = TRUE, ...)
}
\arguments{
\item{x}{modelname}

\item{verbose}{whether to display messages}

\item{project}{path to project directory}

\item{opt}{alternative specification of project}

\item{rundir}{model specific run directory}

\item{ctlfile}{path to model control stream}

\item{ext}{extension (with dot) for control stream}

\item{parse}{convert thetas to initList}

\item{...}{dots}
}
\value{
nmctl
}
\description{
Converts modelname to nmctl.
}

