% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{as.nmctl}
\alias{as.nmctl}
\title{Coerce to NONMEM Control Object}
\usage{
as.nmctl(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{dots}
}
\value{
nmctl
}
\description{
Coerces to NONMEM control stream object.
}

