% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{as.nmctlType.nmctl}
\alias{as.nmctlType.nmctl}
\title{Extract nmctl record type from nmctl}
\usage{
\method{as.nmctlType}{nmctl}(x, type, ...)
}
\arguments{
\item{x}{nmctl}

\item{type}{theta omega or sigma}

\item{...}{dots}
}
\value{
nmctltype (list)
}
\description{
Extracts nmctl record type from nmctl.
}

