% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{as.nmctlType}
\alias{as.nmctlType}
\title{Extract nmctl record type}
\usage{
as.nmctlType(x, ...)
}
\arguments{
\item{x}{nmctl}

\item{...}{dots}
}
\value{
nmctltype (list)
}
\description{
Extracts nmctl record type.
}

