% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{as.paramComments.initList}
\alias{as.paramComments.initList}
\title{Convert initList to paramComments}
\usage{
\method{as.paramComments}{initList}(x, type, prior, ...)
}
\arguments{
\item{x}{initlist}

\item{type}{parameter type: theta, omega,or sigma}

\item{prior}{number of prior parameters of this type (imporant for numbering)}

\item{...}{dots}
}
\value{
data.frame
}
\description{
Converts initList to paramComments
}

