% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{as.paramComments.nmctl}
\alias{as.paramComments.nmctl}
\title{Convert nmctl to paramComments}
\usage{
\method{as.paramComments}{nmctl}(x, fields, expected, na, ...)
}
\arguments{
\item{x}{nmctl}

\item{fields}{data items to scavenge from control stream comments}

\item{expected}{parameters known from NONMEM output}

\item{na}{string to use for NA values when writing default metafile}

\item{...}{dots}
}
\value{
data.frame
}
\description{
Converts nmctl to paramComments
}

