% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{as.paramComments.nmctlType}
\alias{as.paramComments.nmctlType}
\title{Convert nmctlType to paramComments}
\usage{
\method{as.paramComments}{nmctlType}(x, ...)
}
\arguments{
\item{x}{nmctlType}

\item{...}{dots}
}
\value{
data.frame
}
\description{
Converts nmctlType to paramComments
}

