% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partab.R
\name{as.partab.character}
\alias{as.partab.character}
\title{Create Model Parameter Table from Character}
\usage{
\method{as.partab}{character}(x, ...)
}
\arguments{
\item{x}{numeric}

\item{...}{arguments to methods}
}
\description{
Creates a model parameter table from a character string.
}
\details{
Reclassifies x as a modelname and calls as.partab again.
}

