% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partab.R
\name{as.partab.numeric}
\alias{as.partab.numeric}
\title{Create Model Parameter Table from Number.}
\usage{
\method{as.partab}{numeric}(x, ...)
}
\arguments{
\item{x}{numeric}

\item{...}{arguments to methods}
}
\description{
Creates a model parameter table from a number.
}
\details{
Just coerces to character and calls as.partab again.
}

