% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partab.R
\name{as.partab.partab}
\alias{as.partab.partab}
\title{Create Model Parameter Table from partab}
\usage{
\method{as.partab}{partab}(x, ...)
}
\arguments{
\item{x}{partab}

\item{...}{arguments to methods}
}
\description{
Creates a model parameter table from a partab object.
}
\details{
Just returns the object unmodified.
}

