% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath.R
\name{as.xml_document.filepath}
\alias{as.xml_document.filepath}
\title{Create xml_document From Filepath}
\usage{
\method{as.xml_document}{filepath}(x, strip.namespace = TRUE, ...)
}
\arguments{
\item{x}{filepath}

\item{strip.namespace}{whether to strip e.g. nm: from xml elements}

\item{...}{passed arguments}
}
\value{
xml_document
}
\description{
Creates an xml_document from filepath
}

