% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath.R
\name{as.xml_document.modelname}
\alias{as.xml_document.modelname}
\title{Create xml_document From modelname}
\usage{
\method{as.xml_document}{modelname}(x, project = if (is.null(opt)) getwd()
  else opt, opt = getOption("project"), rundir = file.path(project, x),
  file = file.path(rundir, paste0(x, ".xml")), ...)
}
\arguments{
\item{x}{filename}

\item{project}{parent directory of model directories}

\item{opt}{alternative argument for setting project}

\item{rundir}{specific model directory}

\item{file}{actual xml storage location; overrides others if specified directly}

\item{...}{passed arguments}

\item{strip.namespace}{whether to strip e.g. nm: from xml elements}
}
\value{
xml_document
}
\description{
Creates an xml_document from modelname
}

