% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{fixed.init}
\alias{fixed.init}
\title{Check if init is fixed}
\usage{
\method{fixed}{init}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{dots}
}
\value{
logical
}
\description{
Checks if init is fixed.
}

