% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{fixed.initList}
\alias{fixed.initList}
\title{Check if initList is fixed}
\usage{
\method{fixed}{initList}(x, ...)
}
\arguments{
\item{x}{initList}

\item{...}{dots}
}
\value{
logical
}
\description{
Checks if initList is fixed.
}

