% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{tweak.init}
\alias{tweak.init}
\title{Tweak init}
\usage{
\method{tweak}{init}(x, sd = 0.13, digits = 3, ...)
}
\arguments{
\item{x}{init}

\item{sd}{numeric}

\item{digits}{integer}

\item{...}{dots}
}
\value{
init
}
\description{
Tweaks init.
}

