% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.init.R
\name{tweak.initList}
\alias{tweak.initList}
\title{Tweak initList}
\usage{
\method{tweak}{initList}(x, sd = 0.13, digits = 3, ...)
}
\arguments{
\item{x}{initList}

\item{sd}{numeric}

\item{digits}{integer}

\item{...}{dots}
}
\value{
initList
}
\description{
Tweaks initList.
}

