% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath.R
\name{xpath.default}
\alias{xpath.default}
\title{Evaluate xpath Expression in default Context}
\usage{
\method{xpath}{default}(x, ...)
}
\arguments{
\item{x}{xml_document}

\item{...}{passed arguments}
}
\value{
vector
}
\description{
Coerces x to xml_document and evaluates.
}
\seealso{
\code{\link{xpath.xml_document}}
}

