% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath.R
\name{xpath}
\alias{xpath}
\title{Evaluate xpath Expression}
\usage{
xpath(x, ...)
}
\arguments{
\item{x}{xml_document}

\item{...}{passed arguments}
}
\description{
Evaluates an xpath expression..
}
\details{
The resulting nodeset is scavenged for text, and coerced to best of numeric or character.
}
\examples{
\dontrun{
options('project') <- 'model'
1044 \%>\% xpath('//etashrink/row/col')
}
}
\seealso{
\code{\link{xpath.xml_document}}
}

