% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath.R
\name{xpath.xml_document}
\alias{xpath.xml_document}
\title{Evaluate xpath Expression in Document Context}
\usage{
\method{xpath}{xml_document}(x, xpath, ...)
}
\arguments{
\item{x}{xml_document}

\item{xpath}{xpath expression to evaluate}

\item{...}{passed arguments}
}
\value{
vector
}
\description{
Evaluates an xpath expression for a given document.
}
\details{
The resulting nodeset is scavenged for text, and coerced to best of numeric or character.
}
\seealso{
\code{\link{xpath.default}}
}

