% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{as.bootstrap.modelname}
\alias{as.bootstrap.modelname}
\title{Create a Bootstrap Table from Modelname in Project Context}
\usage{
\method{as.bootstrap}{modelname}(x, project = if (is.null(opt)) getwd() else
  opt, opt = getOption("project"), rundir = file.path(project, x),
  pattern = "bootstrap_results.csv", bootcsv = dir(rundir, pattern =
  pattern, recursive = TRUE, full.names = TRUE), ...)
}
\arguments{
\item{x}{object of dispatch}

\item{project}{path to model directories}

\item{opt}{alternative specification of project}

\item{rundir}{model specific run directory}

\item{pattern}{pattern to search for bootstrap file}

\item{bootcsv}{candidate bootstrap file paths}

\item{...}{arguments to methods}
}
\value{
data.frame
}
\description{
Creates a bootstrap table from a modelname in Project Context.
}
\details{
Assumes project has been identified, model directory exists, and PsN bootstrap method has been run for the model.  Scavenges for the last file matching pattern.
}
\seealso{
\code{\link{as.bootstrap.filepath}}
}

