% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{as.bootstrap}
\alias{as.bootstrap}
\title{Create a Bootstrap Table}
\usage{
as.bootstrap(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{arguments to methods}
}
\value{
data.frame
}
\description{
Creates a bootstrap table.
}
\examples{
\dontrun{
options('project') <- 'model'
1044 \%>\% as.bootstrap
}
}
\seealso{
\code{\link{as.bootstrap.modelname}}
}

