% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx.R
\name{as.flextable}
\alias{as.flextable}
\title{Coerce to flextable}
\usage{
as.flextable(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed to methods}
}
\description{
Coerces to flextable.
}
\seealso{
\code{\link{ReporteRs::docx}}

\code{\link{ReporteRs::addFlexTable}}

\code{\link{ReporteRs::FlexTable}}
}

