% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.nmctl.R
\name{as.paramComments}
\alias{as.paramComments}
\title{Coerce to paramComments}
\usage{
as.paramComments(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{dots}
}
\description{
Coerces to paramComments
}

