% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partab.R
\name{as.partab.modelname}
\alias{as.partab.modelname}
\title{Create a Parameter Table from Model name.}
\usage{
\method{as.partab}{modelname}(x, verbose = TRUE, lo = "2.5", hi = "97.5",
  strip.namespace = TRUE, skip = 28, check.names = FALSE, project = if
  (is.null(opt)) getwd() else opt, opt = getOption("project"),
  rundir = file.path(project, x), metafile = file.path(rundir, paste0(x,
  ".meta")), digits = numeric(0), ci = FALSE, open = "(", close = ")",
  sep = ", ", format = ci, fields = c("symbol", "label", "unit"),
  relative = FALSE, percent = relative, quote = FALSE, na = ".", ...)
}
\arguments{
\item{x}{a model name (numeric or character)}

\item{verbose}{set FALSE to suppress messages}

\item{lo}{the PsN bootstrap lower confidence limit (\%)}

\item{hi}{the PsN bootstrap upper confidence limit (\%)}

\item{strip.namespace}{whether to strip e.g. nm: from xml elements for easier xpath syntax}

\item{skip}{number of lines to skip in bootstrap_results.csv}

\item{check.names}{passed to bootstrap reader}

\item{project}{parent directory of model directories}

\item{opt}{alternative argument for setting project}

\item{rundir}{specific model directory}

\item{metafile}{metadata for parameter table; will be created if missing (edit and re-run)}

\item{digits}{limits numerics to significant digits if specified}

\item{ci}{combine bootstrap lo and hi into an enclosed interval}

\item{open}{first character for bootstrap interval}

\item{close}{last character for bootstrap interval}

\item{sep}{separator for bootstrap interval}

\item{format}{format numerics as character}

\item{fields}{metadata fields to read from control stream if no metafile}

\item{relative}{transform standard errors to relative standard errors: rse replaces se}

\item{percent}{if relative is true, express as percent (else ignore): prse replaces se}

\item{quote}{quote csv values when creating template metafile}

\item{na}{how to encode NA in metafiles}

\item{...}{passed to other functions}
}
\value{
data.frame
}
\description{
Creates a parameter table from a model name. Pass the project argument or set 
the project option.
}
\details{
Normally you can just call the generic.
}

