% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partab.R
\name{as.partab}
\alias{as.partab}
\title{Create Model Parameter Table in Project Context.}
\usage{
as.partab(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{arguments to methods}
}
\description{
Creates a model parameter table in project context.
}
\details{
x can be numeric or character model name, assuming project is identified by argument or option.
}
\examples{
library(magrittr)
options(project = system.file('project/model',package='partab'))
1001 \%>\% as.partab
}
\seealso{
\code{\link{as.partab.modelname}}
}

