% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpath.R
\name{as.xml_document}
\alias{as.xml_document}
\title{Create an xml_document in a Project Context}
\usage{
as.xml_document(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{arguments to methods}
}
\value{
xml_document
}
\description{
Creates an xml_document in a project context.
}
\examples{
\dontrun{
options('project') <- 'model'
1044 \%>\% as.xml_document
}
}

